New behaviour: Jump to latest
When GoBackThere command is called, but no position was saved in the
current view (buffer/file), it will try to jump to the latest used
position (used either by SaveBackThere or GoBackThere command) if
at least one position was saved in the current session.

This is, in a way, made in response to feature-request #5.

Signed-off-by: Iulius Curt <iulius.curt@gmail.com>